Extract api client composable
